Move state up to App level
Moving the state up links the get started
and log in to spotify button so when log
out is clicked on either they both update
to the appropriate values before is a user
is authenticated and authorized. Also some
tiny CSS changes were made on the content
Component to make the text look a little
cleaner.